Fixed alignement

+++ Source/Includes/BSP/bios_call.h
@@ -80,6 +80,6 @@ typedef struct bios_int_regs bios_int_regs_t;
  * @param[in] intnum The interrupt number to raise.
  * @param[in] regs The array containing the registers values for the call.
  */
-void bios_call(const uint8_t intnum, bios_int_regs_t* regs);
+void bios_call(const uint32_t intnum, bios_int_regs_t* regs);
 
 #endif /* __BIOS_CALL_H_  */
